bookstore/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==1.11.6
django-redis==4.8.0
django-tinymce==2.6.0
itsdangerous==0.24
jieba==0.39
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
pycryptodome==3.4.7
PyMySQL==0.7.11
python-alipay-sdk==1.7.0
pytz==2017.2
redis==2.10.6
#uWSGI==2.0.15
vine==1.1.4
Whoosh==2.7.4
gunicorn
